+++ game/ConsoleApplication1/ConsoleApplication1/GameConstants.h
@@ -7,9 +7,9 @@
 class GameConstants {
 public:
 	//radius in which players can walk about the centre of the map
-	static int const TOWER_RADIUS; // defined in GameConstants.cpp	
+	static float const TOWER_RADIUS; // defined in GameConstants.cpp	
 
-	static int const WALK_RADIUS; // defined in GameConstants.cpp	
+	static float const WALK_RADIUS; // defined in GameConstants.cpp	
 	static float const PLAYER_RADIUS; // defined in GameConstants.cpp	
 	static float const PLAYER_SPEED; // defined in GameConstants.cpp
